Extract utility function
