+++ src/components/views/rooms/SendMessageComposer.tsx
@@ -57,36 +57,7 @@ import { ComposerType } from "../../../dispatcher/payloads/ComposerInsertPayload
 import { getSlashCommand, isSlashCommand, runSlashCommand, shouldSendAnyway } from "../../../editor/commands";
 import { KeyBindingAction } from "../../../accessibility/KeyboardShortcuts";
 import { PosthogAnalytics } from "../../../PosthogAnalytics";
-import { getNestedReplyText, makeReplyMixIn } from '../../../utils/Reply';
-
-interface IAddReplyOpts {
-    permalinkCreator?: RoomPermalinkCreator;
-    includeLegacyFallback?: boolean;
-    inThread?: boolean;
-}
-
-function addReplyToMessageContent(
-    content: IContent,
-    replyToEvent: MatrixEvent,
-    opts: IAddReplyOpts = {
-        includeLegacyFallback: true,
-    },
-): void {
-    const replyContent = makeReplyMixIn(replyToEvent, opts.inThread);
-    Object.assign(content, replyContent);
-
-    if (opts.includeLegacyFallback) {
-        // Part of Replies fallback support - prepend the text we're sending
-        // with the text we're replying to
-        const nestedReply = getNestedReplyText(replyToEvent, opts.permalinkCreator);
-        if (nestedReply) {
-            if (content.formatted_body) {
-                content.formatted_body = nestedReply.html + content.formatted_body;
-            }
-            content.body = nestedReply.body + content.body;
-        }
-    }
-}
+import { addReplyToMessageContent } from '../../../utils/Reply';
 
 export function attachRelation(
     content: IContent,
